+++ R/methods_brglm2.R
@@ -13,3 +13,16 @@ get_coef.brmultinom <- function(model, ...) {
 #' @export
 get_predict.brmultinom <- get_predict.multinom
 
+
+#' @include get_group_names.R
+#' @include methods_nnet.R
+#' @rdname get_group_names
+#' @export
+get_group_names.bracl <- get_group_names.multinom
+
+
+#' @rdname get_coef
+#' @export
+get_coef.bracl <- function(model, ...) {
+    stats::coef(model)
+}